offers/requirements.txt
base==0.0.0
blinker==1.6.2
certifi==2023.7.22
click==8.1.6
contextlib2==21.6.0
distlib==0.3.7
fields==5.0.0
filelock==3.12.2
Flask==2.3.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
pi==0.1.2
pipenv==2023.7.23
platformdirs==3.10.0
schema==0.7.5
SQLAlchemy==2.0.19
typing_extensions==4.7.1
virtualenv==20.24.3
virtualenv-clone==0.5.7
Werkzeug==2.3.6
zipp==3.16.2

python-dotenv~=1.0.0
requests~=2.31.0
jsonschema~=4.19.0